% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{DropletTestFiles}
\alias{DropletTestFiles}
\title{The \pkg{DropletTestFiles} package}
\description{
The \pkg{DropletTestFiles} package stores files for droplet-based single-cell protocols,
such as those generated from 10X Genomics' CellRanger software.
These files are primarily intended for testing of utilities to handle raw pipeline output and are not (immediately) analysis-ready, 
unlike the SingleCellExperiment-formatted datasets offered by \pkg{scRNAseq} and related packages.
After all, how else would we test the functions that are required to \emph{make} such datasets analysis-ready?
}
\author{
Aaron Lun
}
