% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTestFiles.R
\name{listTestFiles}
\alias{listTestFiles}
\alias{getTestFile}
\title{List available test files}
\usage{
listTestFiles(hub = ExperimentHub())

getTestFile(path, hub = ExperimentHub(), prefix = FALSE)
}
\arguments{
\item{hub}{An \linkS4class{ExperimentHub} object.}

\item{path}{String containing the \code{RDataPath} of the resource of interest.}

\item{prefix}{Logical scalar indicating whether \code{path} should be prefixed with \code{"DropletTestFiles/"}.}
}
\value{
A \linkS4class{DataFrame} containing information on all relevant test files.
}
\description{
List all test files managed by \pkg{DropletTestFiles} on \pkg{ExperimentHub}.
}
\examples{
out <- listTestFiles()
out

\dontrun{
# Downloading a resource to a local file, using an EH number:
ExperimentHub()[[rownames(out)[1]]]

# Downloading a resource to a local file, using an RDataPath:
getTestFile(out$rdatapath[1])
}

}
\author{
Aaron Lun
}
