% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTestFiles.R
\name{listTestFiles}
\alias{listTestFiles}
\alias{getTestFile}
\title{List available test files}
\usage{
listTestFiles(hub = ExperimentHub(), dataset = NULL, version = NULL)

getTestFile(path, hub = ExperimentHub(), prefix = FALSE)
}
\arguments{
\item{hub}{An \linkS4class{ExperimentHub} object.}

\item{dataset}{String specifying the dataset to search for, e.g., \code{"tenx-2.1.0-pbmc4k"}.
By default, files are listed for all datasets.}

\item{version}{String specifying the version of the datasets to search for.
By default, files are listed for all versions.}

\item{path}{String containing the \code{RDataPath} of the resource of interest.}

\item{prefix}{Logical scalar indicating whether \code{path} should be prefixed with \code{"DropletTestFiles/"}.}
}
\value{
A \linkS4class{DataFrame} containing information on all relevant test files.
}
\description{
List all test files managed by \pkg{DropletTestFiles} on \pkg{ExperimentHub}.
}
\details{
Versions refer to those of the \pkg{DropletTestFiles} package itself,
and refer to the state of the dataset at that version of the package.
}
\examples{
out <- listTestFiles()
out

\dontrun{
# Downloading a resource to a local file, using an EH number:
ExperimentHub()[[rownames(out)[1]]]

# Downloading a resource to a local file, using an RDataPath:
getTestFile(out$rdatapath[1])
}

}
\author{
Aaron Lun
}
